Golang/2/rest/main_test.go
package main

import (
	"app/gorm"
	"bytes"
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/gin-gonic/gin"
	"github.com/stretchr/testify/assert"
)

func setupRouter() *gin.Engine {
	r := gin.Default()

	gorm.Connect()

	r.GET("/users", GetUsers)
	r.POST("/users", CreateUser)
	r.PUT("/users/:id", UpdateUser)
	r.DELETE("/users/:id", DeleteUser)

	return r
}

func TestGetUsers(t *testing.T) {
	r := setupRouter()

	req, _ := http.NewRequest("GET", "/users?page=1&limit=10", nil)
	w := httptest.NewRecorder()
	r.ServeHTTP(w, req)

	assert.Equal(t, http.StatusOK, w.Code)
}

func TestCreateUser(t *testing.T) {
	r := setupRouter()

	userJSON := `{"name": "Test User", "age": 30}`
	req, _ := http.NewRequest("POST", "/users", bytes.NewBuffer([]byte(userJSON)))
	w := httptest.NewRecorder()
	r.ServeHTTP(w, req)

	assert.Equal(t, http.StatusOK, w.Code)
}

func TestUpdateUser(t *testing.T) {
	r := setupRouter()

	userJSON := `{"name": "Updated User", "age": 25}`
	req, _ := http.NewRequest("PUT", "/users/136", bytes.NewBuffer([]byte(userJSON)))
	w := httptest.NewRecorder()
	r.ServeHTTP(w, req)

	assert.Equal(t, http.StatusOK, w.Code)
}

func TestDeleteUser(t *testing.T) {
	r := setupRouter()

	req, _ := http.NewRequest("DELETE", "/users/136", nil)
	w := httptest.NewRecorder()
	r.ServeHTTP(w, req)

	assert.Equal(t, http.StatusOK, w.Code)
}
